+++ tests/lib/async/stream_periodic_test.dart
@@ -16,7 +16,10 @@ main() {
     subscription = stream.listen(expectAsync((data) {
       expect(data, isNull);
       receivedCount++;
-      if (receivedCount == 5) subscription.cancel();
+      if (receivedCount == 5) {
+        var future = subscription.cancel();
+        expect(future, completes);
+      }
     }, count: 5));
   });
 }