+++ CHANGELOG.md
@@ -7,6 +7,12 @@
     on Mac. Was already non-blocking on all other platforms.
   * Report a better error when a bind fails because of a bad source address.
 
+* `dart:async`
+  * More aggressively returns a Future on Stream.cancel operations.
+    Discourages to return `null` from `cancel`.
+  * Fixes a few bugs where the cancel future wasn't passed through
+    transformations.
+
 ### Tool Changes
 
 * `dartfmt` - upgraded to v0.2.9